Turned hardcode into dynamic components
Learnt to generate jsx and store into a variable before passing the jsx into the return statement of components, because if you want to return a mapped jsx, it gets messy and hard to debug. Next step is to make the functional component interact with the states in the original class components.