added updating price feature
Not fully functional, as of now the price state is always set to the price of the option that was previously clicked. I guess this has something to do with state not being updated immediately after the click. (because I've combined resetting responses state function call together with the function call that calculates the total price)